refactor: 쿠폰 도메인의 패키지 위치 변경, 리터럴 상수화 (#134)
- CreateCopuonJobConfig, CouponGroupModel 클래스의 패키지 이동
- 배치 작업의 작업명, 쿼리 리터럴 상수화
