+++ src/test/java/com/firstone/greenjangteo/coupon/testutil/CouponTestObjectFactory.java
@@ -1,8 +1,8 @@
 package com.firstone.greenjangteo.coupon.testutil;
 
+import com.firstone.greenjangteo.application.model.CouponGroupModel;
 import com.firstone.greenjangteo.coupon.dto.IssueCouponsRequestDto;
 import com.firstone.greenjangteo.coupon.model.Amount;
-import com.firstone.greenjangteo.coupon.model.CouponGroupModel;
 import com.firstone.greenjangteo.coupon.model.ExpirationPeriod;
 import com.firstone.greenjangteo.coupon.model.IssueQuantity;
 import com.firstone.greenjangteo.coupon.model.entity.CouponGroup;
